media-agent/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
langchain
langchain-openai
langchain-core
langgraph
google-cloud-bigquery
google-auth
pytest
pytest-asyncio
httpx